Move tests to a more logical place
